refactor(*): Move config server integration to kork-config
