+++ clouddriver-kubernetes/clouddriver-kubernetes.gradle
@@ -16,6 +16,7 @@ dependencies {
   implementation "com.netflix.spinnaker.fiat:fiat-api:$fiatVersion"
   implementation "com.netflix.spinnaker.fiat:fiat-core:$fiatVersion"
   implementation "com.netflix.spinnaker.kork:kork-artifacts"
+  implementation "com.netflix.spinnaker.kork:kork-config"
   implementation "com.netflix.spinnaker.moniker:moniker"
   implementation "io.fabric8:kubernetes-client:4.1.1"
   implementation "io.kubernetes:client-java:5.0.0"
