+++ clouddriver-web/src/main/groovy/com/netflix/spinnaker/clouddriver/Main.groovy
@@ -18,6 +18,8 @@ package com.netflix.spinnaker.clouddriver
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.netflix.spinnaker.clouddriver.security.config.SecurityConfig
+import com.netflix.spinnaker.kork.boot.DefaultPropertiesBuilder
+import com.netflix.spinnaker.kork.configserver.ConfigServerBootstrap
 import org.springframework.boot.actuate.autoconfigure.elasticsearch.ElasticSearchJestHealthIndicatorAutoConfiguration
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration
 import org.springframework.boot.autoconfigure.batch.BatchAutoConfiguration
@@ -59,26 +61,7 @@ import java.security.Security
 ])
 @EnableScheduling
 class Main extends SpringBootServletInitializer {
-
-  private static final Map<String, String> DEFAULT_PROPS = [
-    'netflix.environment'     : 'test',
-    'netflix.account'         : '${netflix.environment}',
-    'netflix.stack'           : 'test',
-    'spring.config.additional-location' : '${user.home}/.spinnaker/',
-    'spring.profiles.active'  : '${netflix.environment},local',
-    // add the Spring Cloud Config "composite" profile to default to a configuration
-    // source that won't prevent app startup if custom configuration is not provided
-    'spring.profiles.include' : 'composite',
-    'spring.config.name'      : 'spinnaker,${spring.application.name}'
-  ]
-
-  private static final Map<String, String> BOOTSTRAP_SYSTEM_PROPS = [
-    'spring.application.name'               : 'clouddriver',
-    // default locations must be included pending the resolution of https://github.com/spring-cloud/spring-cloud-commons/issues/466
-    'spring.cloud.bootstrap.location'       : 'classpath:/,classpath:/config/,file:./,file:./config/,${user.home}/.spinnaker/',
-    'spring.cloud.bootstrap.name'           : 'spinnakerconfig,${spring.application.name}config',
-    'spring.cloud.config.server.bootstrap'  : 'true'
-  ]
+  private static final Map<String, Object> DEFAULT_PROPS = new DefaultPropertiesBuilder().build()
 
   static {
     /**
@@ -91,8 +74,7 @@ class Main extends SpringBootServletInitializer {
   }
 
   static void main(String... args) {
-    BOOTSTRAP_SYSTEM_PROPS.findAll { key, value -> !System.getProperty(key)}
-      .each { key, value -> System.setProperty(key, value)}
+    ConfigServerBootstrap.systemProperties("clouddriver")
     new SpringApplicationBuilder()
       .properties(DEFAULT_PROPS)
       .sources(Main)
